replaced server component with generic http-server
