+++ how-to.v1/use-notifications-proxy/platform-notification-proxy/client/tsconfig.json
@@ -10,7 +10,7 @@
 		"esModuleInterop": true,
 		"forceConsistentCasingInFileNames": true,
 		"moduleResolution": "node",
-		"types": ["../common/types/fin"]
+		"types": ["./fin"]
 	},
-	"include": ["./src/**/*.ts"]
+	"include": ["./src/**/*.ts", "fin.d.ts"]
 }
