--- how-to.v1/use-notifications-proxy/platform-notification-proxy/common/LICENSE.MD
@@ -1,6 +1,0 @@
->**_:information_source: OpenFin Workspace:_** [OpenFin Workspace](https://www.openfin.co/workspace/) is a commercial product and this repo is for evaluation purposes. Use of the OpenFin Container and OpenFin Workspace components is only granted pursuant to a  license from OpenFin. Please [**contact us**](https://www.openfin.co/workspace/poc/) if you would like to request a developer evaluation key or to discuss a production license.
-OpenFin Workspace is currently **only supported on Windows** although you can run the sample on a Mac for development purposes.
-
-The developer license can be found here:
-
-* [https://www.openfin.co/developer-agreement/](https://www.openfin.co/developer-agreement/)
