+++ how-to.v1/use-notifications-proxy/platform-notification-proxy/public/style/style-view.json
