+++ how-to.v1/use-notifications-proxy/platform-notification-proxy/scripts/kill.mjs
