--- how-to.v2/use-notifications-proxy/platform-notification-proxy/common/package.json
@@ -1,44 +1,0 @@
-{
-	"name": "openfin-workspace--common",
-	"version": "12.6.0",
-	"description": "Used to provide a common directory to host assets that apply across examples",
-	"main": "index.js",
-	"scripts": {
-		"dos": "desktop-owner-settings.bat && (npm run kill)",
-		"kill": "node ./scripts/kill.mjs",
-		"client": "node ./scripts/launch.mjs",
-		"client-watch": "node ./scripts/launch.mjs http://localhost:8080/manifest.fin.json ./public/*.fin.json",
-		"secondclient": "node ./scripts/launch.mjs http://localhost:8080/second.manifest.fin.json",
-		"thirdclient": "node ./scripts/launch.mjs http://localhost:8080/third.manifest.fin.json",
-		"fourthclient": "node ./scripts/launch.mjs http://localhost:8080/fourth.manifest.fin.json",
-		"build": "npm run build-client",
-		"build-client": "webpack build --config ./client/webpack.config.js --mode=development",
-		"build-client-prod": "webpack build --config ./client/webpack.config.js --mode=production",
-		"setup": "cd ../../ && npm install && cd how-to/common && npm run build",
-		"validate": "npm --prefix ../../ run validate"
-	},
-	"author": "john.mandia@openfin.co",
-	"contributors": [
-		"adam.saland@openfin.co",
-		"martyn.janes@openfin.co",
-		"pj.singh@openfin.co"
-	],
-	"license": "SEE LICENSE IN LICENSE.MD",
-	"dependencies": {
-		"@openfin/workspace": "12.6.6",
-		"@openfin/workspace-platform": "12.6.6",
-		"csstype": "^3.1.2"
-	},
-	"devDependencies": {
-		"@openfin/core": "30.74.13",
-		"@types/express": "^4.17.17",
-		"@types/node": "^20.1.3",
-		"chokidar": "^3.5.3",
-		"express": "^4.18.2",
-		"openfin-adapter": "30.74.13",
-		"ts-loader": "^9.4.2",
-		"typescript": "^4.9.5",
-		"webpack": "^5.82.1",
-		"webpack-cli": "^5.1.1"
-	}
-}
