+++ how-to.v2/use-notifications-proxy/platform-notification-proxy/public/platform/provider.html
@@ -5,7 +5,7 @@
 		<title>Notifications Proxy</title>
 		<meta name="description" content="" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<link rel="stylesheet" href="/common/style/app.css" />
+		<link rel="stylesheet" href="/style/app.css" />
 		<script src="../js/provider.bundle.js"></script>
 	</head>
