+++ how-to.v2/use-notifications-proxy/platform-notification-proxy/public/style/style-view.json
