+++ how-to.v2/use-notifications-proxy/platform-notification-proxy/scripts/kill.mjs
