--- how-to.v2/use-notifications-proxy/platform-notification-proxy/server/src/index.ts
@@ -1,17 +1,0 @@
-import express from "express";
-import path from "path";
-
-const app = express();
-
-const mainPath = path.join(__dirname, "..", "..", "public");
-const commonPath = path.join(__dirname, "..", "..", "common", "public");
-const port = 8080;
-
-console.log(`Main Path / = ${mainPath}`);
-console.log(`Common Path /common = ${commonPath}`);
-console.log("Root path checks both the main and common directories.");
-app.use(express.static(mainPath));
-app.use("/common", express.static(commonPath));
-app.listen(port, () => {
-	console.log("server is listening on port", port);
-});
